requirements.txt
matplotlib==3.7.1
numpy==1.23.5
pandas==2.0.0
scikit_learn==1.2.2
tensorflow_macos==2.12.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
